requirements.txt
easydict==1.13
hdbscan==0.8.40
matplotlib==3.7.4
numpy==1.24.4
open3d==0.16.0
paramiko==3.5.0
Pillow==10.1.0
pynput==1.7.7
PyYAML==6.0.2
scikit_learn==1.3.2
tqdm==4.66.4
